+++ BreakInAndOut/src/UI/levelPanel.cpp
@@ -6,8 +6,8 @@
 
 void UI::levelPanel(Scene* scene) {
 
-	const float screenWidth = GetScreenWidth();
-	const float screenHeight = GetScreenHeight();
+	const float screenWidth = static_cast<float>(GetScreenWidth());
+	const float screenHeight = static_cast<float>(GetScreenHeight());
 
 	const Vector2 panePosition = { screenWidth  - 400.0f, 0.0f };
 	const Vector2 panePadding = { 50.0f, 50.0f };
@@ -19,7 +19,7 @@ void UI::levelPanel(Scene* scene) {
 		const Vector2 position = { panePosition.x + panePadding.x, panePosition.y + panePadding.y + yPosition};
 		const std::string text = "Level:";
 		const Color color = BLACK;
-		const float fontSize = 30;
+		const int fontSize = 30;
 		Entity::createText(scene, position, text, color, fontSize);
 		yPosition += fontSize;
 	}
@@ -29,7 +29,7 @@ void UI::levelPanel(Scene* scene) {
 		const Vector2 position = { panePosition.x + panePadding.x, panePosition.y + panePadding.y + yPosition};
 		const std::string text = scene->level->getName();
 		const Color color = BLACK;
-		const float fontSize = 30;
+		const int fontSize = 30;
 		Entity::createText(scene, position, text, color, fontSize);
 		yPosition += fontSize;
 	}
@@ -40,7 +40,7 @@ void UI::levelPanel(Scene* scene) {
 		const Vector2 position = { panePosition.x + panePadding.x, panePosition.y + panePadding.y + yPosition};
 		const std::string text = "Layer:";
 		const Color color = BLACK;
-		const float fontSize = 20;
+		const int fontSize = 20;
 		Entity::createText(scene, position, text, color, fontSize);
 		yPosition += fontSize;
 	}
@@ -73,7 +73,7 @@ void UI::levelPanel(Scene* scene) {
 		const Vector2 position = { panePosition.x + panePadding.x, panePosition.y + panePadding.y + yPosition};
 		const std::string text = "Remaining:";
 		const Color color = BLACK;
-		const float fontSize = 20;
+		const int fontSize = 20;
 		Entity::createText(scene, position, text, color, fontSize);
 		yPosition += fontSize;
 	}
@@ -82,7 +82,7 @@ void UI::levelPanel(Scene* scene) {
 		const Vector2 position = { panePosition.x + panePadding.x, panePosition.y + panePadding.y + yPosition};
 		const std::string text = "0";
 		const Color color = BLACK;
-		const float fontSize = 100;
+		const int fontSize = 100;
 		Entity::Entity displayText = Entity::createText(scene, position, text, color, fontSize);
 		displayText.addComponent<Component::Watcher>(Component::Watcher{
 			[scene, displayText]() {
